# crates/tornas/src/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "tornas API",
    "version": "0.1.0",
    "description": "Home media center: torrent-backed movie library with an LRU disk budget. Movies are keyed by IMDb id. No authentication; intended for a home LAN."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/movies": {
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "List movies",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "checking",
                "downloading",
                "seeding",
                "done",
                "paused",
                "error",
                "missing"
              ]
            },
            "description": "Only movies in this state"
          }
        ],
        "responses": {
          "200": {
            "description": "Movies, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "movies"
        ],
        "summary": "Add a movie",
        "description": "Looks the IMDb id up on TMDB, resolves the magnet, evicts least-recently-used movies until the video fits the disk budget, then starts the download. Returns once the download has started.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMovie"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "URL of the new movie"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "description": "TMDB has no movie for that IMDb id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Movie already in the library, or everything is paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "description": "Validation failed (bad imdb_id, magnet or JSON shape)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "TMDB unreachable or returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "507": {
            "description": "Cannot free enough space: everything evictable is protected, or the file exceeds the whole budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/movies/{imdb_id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ImdbId"
        }
      ],
      "get": {
        "tags": [
          "movies"
        ],
        "summary": "Get a movie",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "tags": [
          "movies"
        ],
        "summary": "Update a movie",
        "description": "Set any of the fields; at least one is required and unknown fields are refused. `last_used_at` set to \"now\" moves the movie to the back of the eviction queue. Changing a limit reloads the torrent in the engine; its downloaded pieces are kept, so nothing is checked or downloaded again.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "last_used_at": {
                    "oneOf": [
                      {
                        "type": "integer",
                        "description": "unix seconds"
                      },
                      {
                        "type": "string",
                        "enum": [
                          "now"
                        ]
                      }
                    ]
                  },
                  "download_limit": {
                    "oneOf": [
                      {
                        "type": "integer",
                        "minimum": 1,
                        "description": "bytes per second"
                      },
                      {
                        "type": "string",
                        "example": "2M",
                        "description": "size per second, like 500K or 2M"
                      }
                    ],
                    "nullable": true,
                    "description": "null removes the movie's own limit"
                  },
                  "upload_limit": {
                    "oneOf": [
                      {
                        "type": "integer",
                        "minimum": 1,
                        "description": "bytes per second"
                      },
                      {
                        "type": "string",
                        "example": "2M",
                        "description": "size per second, like 500K or 2M"
                      }
                    ],
                    "nullable": true,
                    "description": "null removes the movie's own limit"
                  },
                  "peer_limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10000,
                    "nullable": true,
                    "description": "peers for this torrent; null goes back to the default"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated movie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Movie"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "movies"
        ],
        "summary": "Remove a movie and delete its files",
        "responses": {
          "204": {
            "description": "Removed"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Recent events (adds, evictions, completions)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 500
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/pause": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Global pause state",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pause"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Pause everything",
        "description": "Stops all downloads, uploads and tracker traffic and refuses new movies. Body optional: omitted uses the configured default (3h). Repeating it while paused moves the end time. Persisted across restarts.",
        "security": [
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "duration": {
                    "oneOf": [
                      {
                        "type": "integer",
                        "description": "seconds"
                      },
                      {
                        "type": "string",
                        "example": "3h"
                      }
                    ]
                  },
                  "indefinite": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pause"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "tags": [
          "system"
        ],
        "summary": "Resume",
        "description": "Idempotent. Finished movies stay paused unless --keep-seeding.",
        "security": [
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pause lifted; includes resumed count"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "ImdbId": {
        "name": "imdb_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^tt[0-9]+$"
        },
        "example": "tt0111161"
      }
    },
    "responses": {
      "Error": {
        "description": "Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "not_found",
                  "conflict",
                  "invalid",
                  "no_space",
                  "upstream",
                  "bad_request",
                  "unsupported_media_type",
                  "internal",
                  "unauthorized"
                ]
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "kind",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "AddMovie": {
        "type": "object",
        "required": [
          "imdb_id"
        ],
        "properties": {
          "imdb_id": {
            "type": "string",
            "pattern": "^tt[0-9]+$",
            "example": "tt0111161"
          },
          "magnet": {
            "type": "string",
            "example": "magnet:?xt=urn:btih:..."
          },
          "initial_peers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "host:port peers to try first; only needed without DHT/trackers",
            "example": [
              "192.168.1.20:6881"
            ]
          },
          "torrent_url": {
            "type": "string",
            "description": "URL of a .torrent file"
          },
          "torrent_base64": {
            "type": "string",
            "description": "Base64-encoded .torrent file"
          }
        },
        "description": "Exactly one of magnet, torrent_url, torrent_base64 is required. Private torrents must be added from their .torrent file, never a magnet, so no DHT lookup happens."
      },
      "Torrent": {
        "type": "object",
        "properties": {
          "info_hash": {
            "type": "string"
          },
          "imdb_id": {
            "type": "string"
          },
          "magnet": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "video_file_idx": {
            "type": "integer"
          },
          "video_file_name": {
            "type": "string"
          },
          "added_at": {
            "type": "integer"
          },
          "download_limit": {
            "type": "integer",
            "nullable": true
          },
          "upload_limit": {
            "type": "integer",
            "nullable": true
          },
          "peer_limit": {
            "type": "integer",
            "nullable": true
          },
          "private": {
            "type": "boolean",
            "description": "BEP 27 private torrent"
          }
        }
      },
      "Movie": {
        "type": "object",
        "properties": {
          "imdb_id": {
            "type": "string"
          },
          "tmdb_id": {
            "type": "integer",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": "integer",
            "nullable": true
          },
          "overview": {
            "type": "string",
            "nullable": true
          },
          "poster_url": {
            "type": "string",
            "nullable": true
          },
          "backdrop_url": {
            "type": "string",
            "nullable": true
          },
          "runtime_min": {
            "type": "integer",
            "nullable": true
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "number",
            "nullable": true
          },
          "added_at": {
            "type": "integer",
            "description": "unix seconds"
          },
          "last_used_at": {
            "type": "integer",
            "description": "unix seconds; drives LRU eviction"
          },
          "torrent": {
            "$ref": "#/components/schemas/Torrent"
          },
          "state": {
            "type": "string",
            "enum": [
              "checking",
              "downloading",
              "seeding",
              "done",
              "paused",
              "error",
              "missing",
              "queued"
            ]
          },
          "progress_bytes": {
            "type": "integer"
          },
          "total_bytes": {
            "type": "integer"
          },
          "finished": {
            "type": "boolean"
          },
          "download_bps": {
            "type": "integer"
          },
          "upload_bps": {
            "type": "integer"
          },
          "peers": {
            "type": "integer"
          },
          "protected": {
            "type": "boolean",
            "description": "inside the stream grace window, so not evictable"
          }
        }
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "ts": {
            "type": "integer"
          },
          "kind": {
            "type": "string",
            "enum": [
              "start",
              "add",
              "evict",
              "remove",
              "done"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Pause": {
        "type": "object",
        "properties": {
          "paused": {
            "type": "boolean"
          },
          "since": {
            "type": "integer",
            "nullable": true
          },
          "until": {
            "type": "integer",
            "nullable": true,
            "description": "unix seconds; null when indefinite"
          },
          "remaining_secs": {
            "type": "integer",
            "nullable": true
          },
          "indefinite": {
            "type": "boolean"
          },
          "default_duration_secs": {
            "type": "integer"
          },
          "reason": {
            "type": "string",
            "enum": [
              "manual",
              "disk_missing"
            ],
            "nullable": true,
            "description": "why everything is paused; a disk_missing pause lifts when the disk returns and cannot be resumed by hand"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "TORNAS_API_TOKEN; required for POST/PATCH/DELETE under /api and for /api/logs when configured"
      }
    }
  }
}